Fix helloworld circular dependency & envAddresses
