+++ typescript/infra/src/config/helloworld/types.ts
@@ -1,16 +1,6 @@
 import { AgentConnectionType, ChainMap, ChainName } from '@hyperlane-xyz/sdk';
-import { MultiProvider, RouterConfig } from '@hyperlane-xyz/sdk';
-import { objMap } from '@hyperlane-xyz/utils';
 
-import { Contexts } from '../../config/contexts';
-import {
-  mainnetHyperlaneDefaultIsmCache,
-  routingIsm,
-} from '../../config/routingIsm';
-import { getRouterConfig } from '../../scripts/utils';
-
-import { DockerConfig } from './agent';
-import { DeployEnvironment } from './environment';
+import { DockerConfig } from '../agent';
 
 export enum HelloWorldKathyRunMode {
   // Sends messages between all pairwise chains
@@ -47,20 +37,4 @@ export interface HelloWorldKathyConfig {
 export interface HelloWorldConfig {
   addresses: ChainMap<{ router: string }>;
   kathy: HelloWorldKathyConfig;
-}
-
-export async function helloWorldRouterConfig(
-  environment: DeployEnvironment,
-  context: Contexts,
-  multiProvider: MultiProvider,
-): Promise<ChainMap<RouterConfig>> {
-  const routerConfig = await getRouterConfig(environment, multiProvider, true);
-  return objMap(routerConfig, (chain, config) => ({
-    ...config,
-    interchainSecurityModule:
-      context === Contexts.Hyperlane
-        ? // TODO move back to `undefined` after these are verified and made the default ISMs
-          mainnetHyperlaneDefaultIsmCache[chain]
-        : routingIsm(environment, chain, context),
-  }));
 }
