+++ typescript/infra/src/config/index.ts
@@ -5,6 +5,6 @@ export {
   StorageGasOracleConfig,
   getAllStorageGasOracleConfigs,
 } from './gas-oracle';
-export { HelloWorldConfig } from './helloworld';
+export { HelloWorldConfig } from './helloworld/types';
 export { InfrastructureConfig } from './infrastructure';
 export * from './agent';
